Create Multi-programming_with_Fixed_Tasks.cpp
Multi-programming with Fixed Number of Tasks (MFT) :
The biggest problem with PCP was that processor cycles were wasted whilst the application program made a request for I/O. To improve processor utilization, another OS called Multi-programming with Fixed number of Tasks (MFT) was introduced.
Multi-programming implies that a number of programs are resident in memory at any given point of time. Of these, one program is running on a processor. When a program makes a request for I/O its stateis saved and the program is suspended. Control is given to another program and it continues from where it left off. In this fashion, ultimately all programs run to completion after many cycles of running and sleeping.In MFT, many fixed partitions were created at machine startup. The.The sizes of the partitions could
vary.Each program would be loaded into a free partition of appropriate size.
MFT was intended to serve as a stop-gap until Multi-programming with a Variable number of Tasks (MVT), the intended "target" configuration of OS/360, became available in 1967. Early versions of MVT had many problems, so the simpler MFT continued to be used for many years. After introducing new System/370 machines with virtual memory, in 1972 MFT was developed into OS/VS1, the last system of this particular line.
The first version of MFT shared much of the code and architecture with PCP, and was limited to four partitions. It was very cumbersome to run multiple partitions. Many installations used Houston Automatic Spooling Priority (HASP) to mitigate the complexity.
MFT Version II (MFT-II)  shared much more of the Control Program and Scheduler code with MVT, and was much more flexible to run. The maximum number of partitions increased to 52.
Later modifications of MFT-II added sub-tasking, so that the fixed number of tasks was no longer fixed, although the number of partitions did remain a limitation.
Experience indicated that it was not advisable to install MFT on systems with less than 256 KB of memory, which in the 1960s was quite a large amount.
Also referred to as SYS=INT in macro expansions that were system-dependent.