changed list length to reflect contents of the list
I changed the return value of n in List_length to reflect the number of items inside the list, so a newly initialized list will return a length of 0.  To prevent items in List_toArray from being cut off, I addeone back to n at the beginning of the List_toArray function.